+++ back/src/main/java/com/udteam/miristock/repository/CommentRepository.java
@@ -3,7 +3,7 @@ package com.udteam.miristock.repository;
 import com.udteam.miristock.entity.CommentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
-    CommentEntity findByArticle_ArticleNo (Long articleNo);
+public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+    CommentEntity findByArticle_ArticleNo (Integer articleNo);
 }
